fix: solve state problems
