+++ webapp/src/components/FilterHome/index.js
@@ -1,0 +1,1 @@
+export { default } from './FilterHome'
